+++ R/survey_module.R
@@ -1,6 +1,18 @@
-#' Create R6-based Shiny modules for the app
+#' Create UI of your questionnaire
 #'
-#' @name survey_module
+#' @import shiny
+#' @export
+#'
+
+quetzio_UI <- function(module_id) {
+  ns <- NS(module_id)
+
+  uiOutput(ns("quetzio_UI"))
+
+}
+
+#' Create R6-based server module to generate and hold the state of your
+#' questionnaire
 #'
 #' @import R6
 #' @import googlesheets4
@@ -11,8 +23,8 @@
 #' @export
 #'
 
-survey_module <- R6::R6Class(
-  "shiny.survey_module",
+quetzio_server <- R6::R6Class(
+  "quetzio_server",
 
   private = list(
 
@@ -21,7 +33,33 @@ survey_module <- R6::R6Class(
     output_gsheet = NULL,
     output_ss = NULL,
     output_sheet = NULL,
-    css = NULL
+    css = NULL,
+    render_ui = NULL,
+    server = function() {
+
+      output <- .survey_backend(
+        id = self$module_id,
+        source_list = self$source_list,
+        mandatory_items = private$mandatory_items,
+        numeric_items = private$numeric_items,
+        output_gsheet = private$output_gsheet,
+        output_ss = private$output_ss,
+        output_sheet = private$output_sheet,
+        button_labels = self$button_labels,
+        div_id = self$div_id,
+        css = private$css,
+        render_ui = private$render_ui
+      )
+
+      observe({
+        req(output)
+
+        self$is_done(reactiveValuesToList(output)$is_done)
+        self$message(reactiveValuesToList(output)$message)
+        self$answers(reactiveValuesToList(output)$answers)
+
+      })
+    }
 
   ),
 
@@ -39,6 +77,25 @@ survey_module <- R6::R6Class(
     #' @field button_labels character vector of length two with labels for submit
     #' button in active and disabled state
     button_labels = NULL,
+
+    #' @field is_done logical reactiveVal indicating if the survey has been completed
+    is_done = NULL,
+
+    #' @field message reactiveVal catching any warning messages
+    message = NULL,
+
+    #' @field answers reactiveValues object containing answers to questions
+    answers = NULL,
+
+    #' @description method to change the state of the UI
+    #'
+    #' @param x logical indicating what should be the state of the UI. If TRUE,
+    #' then the UI will be rendered.
+    toggle_ui = function(x) {
+
+      private$render_ui(isTRUE(x))
+
+    },
 
     #' @description Initializing the shiny.survey_module object
     #'
@@ -67,6 +124,8 @@ survey_module <- R6::R6Class(
     #' If not specified, default look will be used
     #' @param button_labels character vector of length two with labels for submit
     #' button in active and disabled state
+    #' @param render_ui logical indicating if the UI for questionnaire should be
+    #' rendered
     #'
     #' @details
     #'
@@ -83,7 +142,7 @@ survey_module <- R6::R6Class(
     #'   and 'output_gsheet_sheetname'
     #'
     #'
-    #' @return the 'shiny.survey_module' object
+    #' @return the 'survey_module_server' object
 
     initialize = function(
       source_method,
@@ -97,8 +156,10 @@ survey_module <- R6::R6Class(
       module_id = NULL,
       div_id = "form",
       custom_css = NULL,
-      button_labels = c("Submit", "Cannot submit")
+      button_labels = c("Submit", "Cannot submit"),
+      render_ui = TRUE
     ){
+
       # initialize checks
 
       # check if all needed arguments are provided for source methods
@@ -144,7 +205,7 @@ survey_module <- R6::R6Class(
         )
 
         # check df validity
-       .check_source_df(source_df)
+        .check_source_df(source_df)
 
         self$source_list <- .df_to_list(
           source_df = source_df
@@ -198,51 +259,21 @@ survey_module <- R6::R6Class(
       self$button_labels <- button_labels
       private$output_gsheet <- output_gsheet
       if(output_gsheet) {
-      private$output_ss <- if (is.null(output_gsheet_id) || is.na(output_gsheet_id)) source_gsheet_id else output_gsheet_id
-      private$output_sheet <- output_gsheet_sheetname
+        private$output_ss <- if (is.null(output_gsheet_id) || is.na(output_gsheet_id)) source_gsheet_id else output_gsheet_id
+        private$output_sheet <- output_gsheet_sheetname
       }
       private$css <- .null_def(
         custom_css,
         ".mandatory_star { color: red; } .invalid_input { outline: red; outline-style: dashed; outline-offset: 10px; }"
-                               )
-
-      },
-
-    #' @description Method to input the UI of your survey into the shinyApp UI
-    #' @return The whole UI of your survey
-
-    ui = function() {
-
-      .generate_ui(
-        source_list = self$source_list,
-        div_id = self$div_id,
-        module_id = self$module_id,
-        css = private$css,
-        button_label = self$button_labels[1]
       )
 
-    },
+      self$is_done <- reactiveVal(FALSE)
+      self$message <- reactiveVal()
+      self$answers <- reactiveVal()
+      private$render_ui <- reactiveVal(render_ui)
 
-    #' @description Method to input the UI of your survey into the shinyApp UI
-    #' @return The reactiveValues holding
-    #' /itemize{
-    #'   /item{is_done logical vector indicating if the survey has been sent. NA indicates error}
-    #'   /item{message if catches error during survey sending, character string with error message. Otherwise NULL }
-    #'   /item{answers when the survey is done, it contains list item values. Otherwise NULL}
-    #' }
+      private$server()
 
-    server = function() {
-
-      .survey_backend(
-        id = self$module_id,
-        source_list = self$source_list,
-        mandatory_items = private$mandatory_items,
-        numeric_items = private$numeric_items,
-        output_gsheet = private$output_gsheet,
-        output_ss = private$output_ss,
-        output_sheet = private$output_sheet,
-        button_labels = self$button_labels
-      )
     }
   )
 )