+++ NEWS.md
@@ -2,3 +2,9 @@
 
 * First production-ready version.
 * Added a `NEWS.md` file to track changes to the package.
+
+# shiny.survey 0.0.4
+
+* Rebuilding of the whole system for modules
+* Begin rebranding to shiny.quetzio
+* Removal of `survey_module` object and creating `quetzio_server` R6 object and `quetzio_UI` function